Revert "new deployment script added"
